refactor: ActorMetrics is an interface with a noop impl when disabled
(cherry picked from commit f2f9cf425e5862bdf760daf97d1aa4cc95cda557)
